create_ai_pdf.py
from fpdf import FPDF
import os

pdf = FPDF()
pdf.add_page()
pdf.set_font("Arial", size=12)

title = "Transforming into an AI-Enabled End-to-End Learning Platform"
content = """
Since you are already using Python/Django, you have a massive advantage: the Python ecosystem is the home of modern AI.

Here is a breakdown of how AI can transform your platform across three pillars:

1. Managing Learning (The "Smart" Tutor)
   - Adaptive Testing (CAT): The system adjusts difficulty in real-time. If a student masters "Magnetism", the next question is harder. Uses Bayesian optimization or LLMs.
   - Personalized Curriculum Generator: Analyzes past attempts to generate custom schedules (e.g., "3-day Mechanics plan").
   - Dynamic Content Generation: Generates variants of existing questions (changing values/context) so students never run out of material.

2. Organizing Thoughts (Knowledge Management)
   - Semantic Search: Allows searching by concept (e.g., "problems with rods in magnetic fields") rather than just keywords. Uses Vector Embeddings.
   - Automated "Knowledge Graphs": Visualizes the student's mastery of connected concepts (Green=Mastered, Red=Weak).
   - Socratic Doubt Assistant: An embedded chatbot gives hints instead of solutions (e.g., "Have you considered conserving angular momentum?").

3. Revising (Retention & Feedback)
   - "Forensic" Mistake Analysis: Analyzes *why* a specific option was chosen (e.g., "You likely forgot the 1/2 factor in KE").
   - AI Smart Flashcards: Generates review cards based specifically on questions the student got wrong.
   - Spaced Repetition Scheduler: Predicts when a concept is about to be forgotten and inserts review questions accordingly.

Recommendation for Phase 1:
Start with "AI-Powered Solution Explanations". Add a "Explain Why I Was Wrong" button to the results page that sends the question and user's answer to an LLM for personalized feedback.
"""

pdf.set_font("Arial", 'B', 16)
pdf.cell(200, 10, txt=title, ln=1, align='C')

pdf.set_font("Arial", size=11)
pdf.multi_cell(0, 8, content)

output_path = "/Users/dhirendrasingh/Desktop/AI_Learning_Platform_Ideas.pdf"
pdf.output(output_path)
print(f"PDF successfully saved to: {output_path}")
