requirements.txt
django
Pillow
django-nested-admin
pynput
